requirements.txt
bcc==0.18.0
blinker==1.9.0
Brlapi==0.8.3
certifi==2020.6.20
chardet==4.0.0
chrome-gnome-shell==0.0.0
click==8.2.1
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distro==1.7.0
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
flask-cors==6.0.1
gyp==0.1
hidpidaemon==18.4.6
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
itsdangerous==2.2.0
jeepney==0.7.1
Jinja2==3.1.6
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
louis==3.20.0
macaroonbakery==1.3.1
MarkupSafe==3.0.2
more-itertools==8.10.0
mysql-connector-python==9.1.0
netaddr==0.8.0
netifaces==0.11.0
numpy==2.2.6
oauthlib==3.2.0
pop-transition==1.1.2
protobuf==3.12.4
pycairo==1.20.1
pycups==2.0.1
pydbus==0.6.0
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.4.0+ubuntu4
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.1.1
python-gnupg==0.4.8
python-xlib==0.29
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
repolib==2.2.1
repoman==1.4.0
requests==2.25.1
SecretStorage==3.3.1
sessioninstaller==0.0.0
six==1.16.0
systemd-python==234
ubuntu-drivers-common==0.0.0
ufw==0.36.1
urllib3==1.26.5
wadllib==1.3.6
Werkzeug==3.1.3
xdg==5
xkit==0.0.0
zipp==1.0.0
